refactor(message-list): split MessageViewHolder into individual files
