+++ legacy/ui/legacy/src/main/java/com/fsck/k9/ui/messagelist/item/MessageListViewHolder.kt
@@ -1,0 +1,7 @@
+package com.fsck.k9.ui.messagelist.item
+
+import android.view.View
+import androidx.recyclerview.widget.RecyclerView.ViewHolder
+
+sealed class MessageListViewHolder(view: View) : ViewHolder(view)
+
